initial
a). Write a program to concatenate two un-ordered doubly linked lists into one. The input and output of the program is shown as follows
_________________________________________________________________
|  Input:									           |
|	List1 : 11 <-> 12 <-> 13 <-> 14 <-> 15          				           |
|	List2 : 44 <-> 21 <-> 56 <-> 5 				                       |
| Outpu	:	 								           |
|             OutputList : 11 <-> 12 <-> 13 <-> 14 <-> 15 <-> 44 <-> 21 <-> 56 <-> 5   |
|________________________________________________________________|
a). Write a program to concatenate two un-ordered doubly linked lists into one. The input and output of the program is shown as follows
b). Modify the program in (a) to combine two ordered doubly linked lists resulting in single ordered doubly linked lists. The output of the program is shown as follows
_________________________________________________________________
|  Input:									           |
|	List1 : 11 <-> 12 <-> 13 <-> 14 <-> 15          				           |
|	List2 : 44 <-> 21 <-> 56 <-> 5 				                       |
| Outpu	t:	 								           |
|             OutputList : 5 <-> 11 <-> 12 <-> 13 <-> 14 <-> 15 <-> 21 <-> 44 <-> 56   |
|________________________________________________________________|
a). Write a program to concatenate two un-ordered doubly linked lists into one. The input and output of the program is shown as follows
b). Modify the program in (a) to combine two ordered doubly linked lists resulting in single ordered doubly linked lists. The output of the program is shown as follows
c). using doubly linked list. Write a function that returns the sum of these two lists. The sum list is linked list representation of the addition of two input numbers. There are two restrictions to solve this problem i.e. you cannot modify the lists and you are not allowed to use explicit extra space (Resultant list is not explicit) Is it easy to implement? or more difficult?
_____________________________________________________
| Input:								           |
| First List: 1 <-> 2 <-> 3  // represents number 123       	           |
| Second List: 9 <-> 9 <-> 9 //  represents number 999     	           |
|								           |
| outPut:							           |
| Resultant list: 1 <-> 1 <-> 2 <->2 // represents number 1122        |
|____________________________________________________|